src/infrastructure/services/JWTokenService/JWTokenService.ts
import { Injectable } from '@nestjs/common';
import { JwtService } from '@nestjs/jwt';
import { IJWTokenService } from '../../../application/interfaces/services/IJWTokenService';

@Injectable()
export class JWTokenService implements IJWTokenService {
  constructor(private readonly jwtService: JwtService) {}

  async generateToken(userId: number): Promise<string> {
    const payload = { sub: userId };
    return this.jwtService.sign(payload);
  }

  async validateToken(token: string): Promise<any> {
    return this.jwtService.verify(token);
  }
}
